Reorganize, add hiding classes
